update-api-login-backend

+++ backend-expressjs/controllers/api/workspace/column.controller.js
@@ -1,33 +1,29 @@
-const { Board, Column, Card } = require("../../models/index");
+const { Column, Card, Board } = require("../../../models/index");
 const { object, string } = require("yup");
 const { Op } = require("sequelize");
-const BoardTransformer = require("../../transformers/board.transformer");
+const ColumnTransformer = require("../../../transformers/workspace/column.transformer");
 
 module.exports = {
   index: async (req, res) => {
-    const { order = "asc", sort = "id", workspace_id, q } = req.query;
+    const { order = "asc", sort = "id", q, board_id } = req.query;
     const filters = {};
-    if (workspace_id) {
-      filters.workspace_id = workspace_id;
+    if (board_id) {
+      filters.board_id = board_id;
     }
     const options = {
       order: [[sort, order]],
       where: filters,
       include: {
-        model: Column,
-        as: "columns",
-        include: {
-          model: Card,
-          as: "cards",
-        },
+        model: Card,
+        as: "cards",
       },
     };
     const response = {};
     try {
-      const boards = await Board.findAll(options);
+      const columns = await Column.findAll(options);
       response.status = 200;
       response.message = "Success";
-      response.data = new BoardTransformer(boards);
+      response.data = new ColumnTransformer(columns);
     } catch (e) {
       response.status = 500;
       response.message = "Server error";
@@ -39,17 +35,13 @@ module.exports = {
     const { id } = req.params;
     const response = {};
     try {
-      const board = await Board.findByPk(id, {
+      const column = await Column.findByPk(id, {
         include: {
-          model: Column,
-          as: "columns",
-          include: {
-            model: Card,
-            as: "cards",
-          },
+          model: Card,
+          as: "cards",
         },
       });
-      if (!board) {
+      if (!column) {
         Object.assign(response, {
           status: 404,
           message: "Not Found",
@@ -58,7 +50,7 @@ module.exports = {
         Object.assign(response, {
           status: 200,
           message: "Success",
-          data: new BoardTransformer(board),
+          data: new ColumnTransformer(column),
         });
       }
     } catch (e) {
@@ -73,19 +65,44 @@ module.exports = {
     if (req.body.title) {
       rules.title = string().required("Chưa nhập tiêu đề");
     }
-    const schema = object(rules);
 
+    if (req.body.board_id) {
+      rules.board_id = string().required("Chưa có board_id");
+    }
+
+    const schema = object(rules);
     const response = {};
     try {
       const body = await schema.validate(req.body, {
         abortEarly: false,
       });
-      const board = await Board.create(body);
+      const column = await Column.create(body);
+      const board = await Board.findByPk(column.board_id);
+
+      if (!board) {
+        await column.destroy();
+
+        Object.assign(response, {
+          status: 404,
+          message: "Not found board",
+        });
+      }
+      let updatedColumnOrderIds = [];
+
+      if (board.columnOrderIds === null) {
+        updatedColumnOrderIds = [column.id];
+      } else {
+        updatedColumnOrderIds = board.columnOrderIds.concat(column.id);
+      }
+
+      await board.update({
+        columnOrderIds: updatedColumnOrderIds,
+      });
 
       Object.assign(response, {
         status: 201,
         message: "Success",
-        data: new BoardTransformer(board),
+        data: new ColumnTransformer(column),
       });
     } catch (e) {
       const errors = Object.fromEntries(
@@ -124,23 +141,19 @@ module.exports = {
       //     body
       //   );
       // }
-      await Board.update(body, {
+      await Column.update(body, {
         where: { id },
       });
-      const board = await Board.findByPk(id, {
+      const column = await Column.findByPk(id, {
         include: {
-          model: Column,
-          as: "columns",
-          include: {
-            model: Card,
-            as: "cards",
-          },
+          model: Card,
+          as: "cards",
         },
       });
       Object.assign(response, {
         status: 200,
         message: "Success",
-        data: new BoardTransformer(board),
+        data: new ColumnTransformer(column),
       });
     } catch (e) {
       const errors = Object.fromEntries(
@@ -154,70 +167,18 @@ module.exports = {
     }
     res.status(response.status).json(response);
   },
-  moveCard: async (req, res) => {
-    const { id } = req.params;
-    const data = req.body;
-    const response = {};
-    try {
-      const columns = data.columns;
-      const board = await Board.findByPk(id);
-      if (!board) {
-        Object.assign(response, {
-          status: 404,
-          message: "Not Found",
-        });
-      }
-
-      for (const column of columns) {
-        const cards = column.cards;
-        await Column.update(
-          {
-            cardOrderIds: column.cardOrderIds,
-          },
-          { where: { id: column.id } }
-        );
-        for (const card of cards) {
-          await Card.update(
-            { column_id: card.column_id },
-            {
-              where: {
-                id: card.id,
-              },
-            }
-          );
-        }
-      }
-
-      await board.update({ columnOrderIds: data.columnOrderIds });
-
-      Object.assign(response, {
-        status: 200,
-        message: "Success",
-      });
-    } catch (e) {
-      Object.assign(response, {
-        status: 500,
-        message: "Server error",
-      });
-    }
-    res.status(response.status).json(response);
-  },
   delete: async (req, res) => {
     const { id } = req.params;
     try {
-      const columns = await Column.findAll({ where: { board_id: id } });
-
-      for (const column of columns) {
+      const column = await Column.findByPk(id);
+      if (column) {
         await Card.destroy({ where: { column_id: column.id } });
+        await column.destroy();
+        res.status(204).json({
+          status: 204,
+          message: "Success",
+        });
       }
-      await Column.destroy({ where: { board_id: id } });
-
-      await Board.destroy({ where: { id }, force: true });
-
-      res.status(204).json({
-        status: 204,
-        message: "Success",
-      });
     } catch (error) {
       res.status(500).json({
         status: 500,
